Move test to shared_axis_transition
